+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   Display the cause of CP not being cancellable in tooltip (#192)
+-   Display function name for each task in workflow & tasks table (#188)
 
 ### Changed